Move APR and TVL calculations.
